Payment/ravepay_services.py
from django.utils.datetime_safe import datetime
from decouple import config
import requests
from Payment.models import Payment


class RavePayServices:
    def __init__(self, data):
        self.data = data
        self.link = None

    @staticmethod
    def package_price(package):
        if package == 'BASIC':
            return 0
        if package == 'CLASSIC':
            return 1000

        if package == 'PREMIUM':
            return 2000

    @staticmethod
    def get_user(user_id):
        from Account.models import User
        return User.objects.get(id=user_id)

    def create_payment_model(self):
        user = self.get_user(self.data['data']['customer']['id'])
        Payment.objects.create(user=user, status='PAID', tx_ref=self.data['data']['tx_ref'],
                               date_of_payment=datetime.now(), package=self.data['data']['customer']['package'])

    def make_payment(self):
        user = self.get_user(self.data['user_id'])
        url = f"{config('RAVE_PAY_BASE_URL')}/payments"
        headers = {"Authorization": f"Bearer {config('RAVE_PAY_KEY')}"}
        tx_ref = self.create_tx_ref()
        data = {"tx_ref": tx_ref, "amount": self.package_price(self.data['package']), "currency": "NGN",
                "redirect_url": f"{config('RAVE_PAY_REDIRECT_URL')}/{user.id}/{self.data['package']}/{tx_ref}",
                "payment_options": self.data['payment_option'],
                "customer": {
                    "id": user.id, "email": user.email, "name": f"{user.first_name} {user.last_name}",
                    "package": self.data['package']
                }, "customizations": {
                "title": "DateFix Payments",
                "description": f"Payment for the {self.data['package']} package",
                "logo": config('LOGO_URL')
            }}
        result = requests.post(url=url, headers=headers, data=data).json()
        if result['status'] == 'success':
            self.link = result['data']['link']
            return True
        return False

    def create_tx_ref(self):
        import secrets
        tx_ref = secrets.token_hex(16)
        try:
            Payment.objects.get(tx_ref__exact=tx_ref)
            self.create_tx_ref()
        except Payment.DoesNotExist:
            return tx_ref
