requirements.txt
aioredis==1.3.1
aniso8601==7.0.0
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
certifi==2020.6.20
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
dj-database-url==0.5.0
Django==3.0.6
django-filter==2.3.0
djangorestframework==3.11.0
graphene==2.1.8
graphene-django==2.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
msgpack==0.6.2
msgpack-python==0.5.6
Pillow==7.1.2
promise==2.3
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.7
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-decouple==3.3
pytz==2020.1
redis==3.5.0
requests==2.24.0
Rx==1.6.1
service-identity==18.1.0
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
Unidecode==1.1.1
urllib3==1.25.10
whitenoise==5.1.0
zope.interface==5.1.0
